Automatically create blank index files in make.sh
Prior to this commit, users trying to use cataloggen on a new file might
have gotten error messages about the absence of the idx and/or ind
files. This commit adds a statement to make.sh to ensure that all files
exist prior to running xelatex for the first time.